+++ examples/data-folder/src/lib.rs
@@ -1,36 +1,13 @@
 use std::fs;
+use wassel_sdk::http::{IntoResponse, Request, StatusCode, handler};
 
-use wassel_sdk_rust::bindings::{
-    export,
-    exports::wassel::foundation::http_handler::{Guest, IncomingRequest, ResponseOutparam},
-    wasi::http::types::{Headers, OutgoingBody, OutgoingResponse},
-};
-
-struct Plugin;
-
-impl Guest for Plugin {
-    fn handle_request(_request: IncomingRequest, response_out: ResponseOutparam) {
-        match fs::read("file.txt") {
-            Ok(data) => write_response(response_out, 200, Some(&data)),
-            Err(e) => write_response(response_out, 500, Some(e.to_string().as_bytes())),
-        }
+#[handler]
+fn handle_request(_request: Request) -> impl IntoResponse {
+    match fs::read("file.txt") {
+        Ok(data) => (StatusCode::OK, data),
+        Err(e) => (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            e.to_string().as_bytes().to_vec(),
+        ),
     }
 }
-
-fn write_response(out: ResponseOutparam, status: u16, body_bytes: Option<&[u8]>) {
-    let res = OutgoingResponse::new(Headers::new());
-    res.set_status_code(status).unwrap();
-
-    if let Some(bytes) = body_bytes {
-        let body = res.body().unwrap();
-        {
-            let stream = body.write().unwrap();
-            stream.write(bytes).unwrap();
-        }
-        OutgoingBody::finish(body, None).unwrap();
-    }
-
-    ResponseOutparam::set(out, Ok(res));
-}
-
-export!(Plugin);