+++ src/http/request.rs
@@ -1,0 +1,3 @@
+use crate::http::Body;
+
+pub type Request = http::Request<Body>;