+++ src/lib.rs
@@ -1,3 +1,5 @@
+pub mod handler;
+pub mod http;
 pub use bindings::export;
 pub mod bindings;
 pub use wasip2 as wasi;